third_party/WebKit/LayoutTests/inspector/tracing/buffer-usage-expected.txt

Tests that buffer usage update are sent when recording trace events and TimelineLifecycleDelegate methods are properly invoked in the expected order.

TimelineLifecycleDelegate.recordingStarted
TimelineLifecycleDelegate.recordingProgress
TimelineLifecycleDelegate.loadingStarted
TimelineLifecycleDelegate.loadingProgress
TimelineLifecycleDelegate.loadingComplete
